+++ src/protocol/xproto.rs
@@ -7609,9 +7609,9 @@ pub struct ConfigureWindowAux {
 }
 impl ConfigureWindowAux {
     fn try_parse(value: &[u8], value_mask: u16) -> Result<(Self, &[u8]), ParseError> {
-        let switch_expr = u32::from(value_mask);
+        let switch_expr = u16::from(value_mask);
         let mut outer_remaining = value;
-        let x = if switch_expr & u32::from(ConfigWindow::X) != 0 {
+        let x = if switch_expr & u16::from(ConfigWindow::X) != 0 {
             let remaining = outer_remaining;
             let (x, remaining) = i32::try_parse(remaining)?;
             outer_remaining = remaining;
@@ -7619,7 +7619,7 @@ impl ConfigureWindowAux {
         } else {
             None
         };
-        let y = if switch_expr & u32::from(ConfigWindow::Y) != 0 {
+        let y = if switch_expr & u16::from(ConfigWindow::Y) != 0 {
             let remaining = outer_remaining;
             let (y, remaining) = i32::try_parse(remaining)?;
             outer_remaining = remaining;
@@ -7627,7 +7627,7 @@ impl ConfigureWindowAux {
         } else {
             None
         };
-        let width = if switch_expr & u32::from(ConfigWindow::WIDTH) != 0 {
+        let width = if switch_expr & u16::from(ConfigWindow::WIDTH) != 0 {
             let remaining = outer_remaining;
             let (width, remaining) = u32::try_parse(remaining)?;
             outer_remaining = remaining;
@@ -7635,7 +7635,7 @@ impl ConfigureWindowAux {
         } else {
             None
         };
-        let height = if switch_expr & u32::from(ConfigWindow::HEIGHT) != 0 {
+        let height = if switch_expr & u16::from(ConfigWindow::HEIGHT) != 0 {
             let remaining = outer_remaining;
             let (height, remaining) = u32::try_parse(remaining)?;
             outer_remaining = remaining;
@@ -7643,7 +7643,7 @@ impl ConfigureWindowAux {
         } else {
             None
         };
-        let border_width = if switch_expr & u32::from(ConfigWindow::BORDER_WIDTH) != 0 {
+        let border_width = if switch_expr & u16::from(ConfigWindow::BORDER_WIDTH) != 0 {
             let remaining = outer_remaining;
             let (border_width, remaining) = u32::try_parse(remaining)?;
             outer_remaining = remaining;
@@ -7651,7 +7651,7 @@ impl ConfigureWindowAux {
         } else {
             None
         };
-        let sibling = if switch_expr & u32::from(ConfigWindow::SIBLING) != 0 {
+        let sibling = if switch_expr & u16::from(ConfigWindow::SIBLING) != 0 {
             let remaining = outer_remaining;
             let (sibling, remaining) = Window::try_parse(remaining)?;
             outer_remaining = remaining;
@@ -7659,7 +7659,7 @@ impl ConfigureWindowAux {
         } else {
             None
         };
-        let stack_mode = if switch_expr & u32::from(ConfigWindow::STACK_MODE) != 0 {
+        let stack_mode = if switch_expr & u16::from(ConfigWindow::STACK_MODE) != 0 {
             let remaining = outer_remaining;
             let (stack_mode, remaining) = u32::try_parse(remaining)?;
             let stack_mode = stack_mode.into();
@@ -7680,7 +7680,7 @@ impl ConfigureWindowAux {
         result
     }
     fn serialize_into(&self, bytes: &mut Vec<u8>, value_mask: u16) {
-        assert_eq!(self.switch_expr(), u32::from(value_mask), "switch `value_list` has an inconsistent discriminant");
+        assert_eq!(self.switch_expr(), u16::from(value_mask), "switch `value_list` has an inconsistent discriminant");
         if let Some(x) = self.x {
             x.serialize_into(bytes);
         }
@@ -7705,28 +7705,28 @@ impl ConfigureWindowAux {
     }
 }
 impl ConfigureWindowAux {
-    fn switch_expr(&self) -> u32 {
+    fn switch_expr(&self) -> u16 {
         let mut expr_value = 0;
         if self.x.is_some() {
-            expr_value |= u32::from(ConfigWindow::X);
+            expr_value |= u16::from(ConfigWindow::X);
         }
         if self.y.is_some() {
-            expr_value |= u32::from(ConfigWindow::Y);
+            expr_value |= u16::from(ConfigWindow::Y);
         }
         if self.width.is_some() {
-            expr_value |= u32::from(ConfigWindow::WIDTH);
+            expr_value |= u16::from(ConfigWindow::WIDTH);
         }
         if self.height.is_some() {
-            expr_value |= u32::from(ConfigWindow::HEIGHT);
+            expr_value |= u16::from(ConfigWindow::HEIGHT);
         }
         if self.border_width.is_some() {
-            expr_value |= u32::from(ConfigWindow::BORDER_WIDTH);
+            expr_value |= u16::from(ConfigWindow::BORDER_WIDTH);
         }
         if self.sibling.is_some() {
-            expr_value |= u32::from(ConfigWindow::SIBLING);
+            expr_value |= u16::from(ConfigWindow::SIBLING);
         }
         if self.stack_mode.is_some() {
-            expr_value |= u32::from(ConfigWindow::STACK_MODE);
+            expr_value |= u16::from(ConfigWindow::STACK_MODE);
         }
         expr_value
     }
